+++ lib/packages/wikibase/data-model-services/RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 # Wikibase DataModel Services release notes
 
-## Version 3.15.0 (dev)
+## Version 3.15.0 (2019-04-24)
 
 * Added `ItemLookup` implementations
   * `LegacyAdapterItemLookup` which is an adapter to `EntityLookup`